perf: replace Int32Array with Float64Array for numeric safety
Replaces Int32Array with Float64Array in computeTaskMetrics to ensure duration arithmetic does not suffer from silent truncation if values ever become fractional or NaN. Fixes the previous type coercion logic flaw.

+++ app.js
@@ -1370,9 +1370,9 @@ function validateDateRange(startLabel, startValue, endLabel, endValue, errors) {
 }
 
 function computeTaskMetrics() {
-  // ⚡ Bolt: Optimize with standard for loops and Int32Array to avoid JS callback overhead, GC overhead and hash lookups for duration tracking
+  // ⚡ Bolt: Optimize with standard for loops and Float64Array to avoid JS callback overhead, GC overhead and hash lookups for duration tracking
   const tasksLen = state.tasks.length;
-  const durationCache = new Int32Array(tasksLen);
+  const durationCache = new Float64Array(tasksLen);
   let totalDays = 0;
 
   for (let i = 0; i < tasksLen; i++) {
